requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Swagger-UI==3.36.0
gunicorn==23.0.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.1
python-dateutil==2.9.0.post0
python-dotenv
pymysql
